+++ packages/ts-components/src/components/opta/football/fixtures-ticker/OptaFootballFixturesTicker.stories.tsx
@@ -6,6 +6,7 @@ import { OptaFootballFixturesTicker } from './OptaFootballFixturesTicker';
 const competitons = [
   '1',
   '2',
+  '3',
   '5',
   '6',
   '8',