+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.47](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.6.46...@times-components/error-view@2.6.47) (2024-05-24)
+
+**Note:** Version bump only for package @times-components/error-view
+
+
+
+
+
 ## [2.6.46](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.6.45...@times-components/error-view@2.6.46) (2024-05-24)
 
 **Note:** Version bump only for package @times-components/error-view