+++ packages/star-button/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.70](https://github.com/newsuk/times-components/compare/@times-components/star-button@0.4.69...@times-components/star-button@0.4.70) (2024-05-24)
+
+**Note:** Version bump only for package @times-components/star-button
+
+
+
+
+
 ## [0.4.69](https://github.com/newsuk/times-components/compare/@times-components/star-button@0.4.68...@times-components/star-button@0.4.69) (2024-05-24)
 
 **Note:** Version bump only for package @times-components/star-button